apiReconhecimento/requirements.txt
fastapi
uvicorn[standard]
deepface
numpy
python-dotenv
mysql-connector-python
requests
